Bug 1899266 - Make ScaleOffset::(un)map_rect operate on min and max points rather than origin and size. r?#gfx-reviewers
ScaleOffset::(un)map_rect() is currently implemented by transforming
the origin and size of each rect. Due to floating point inaccuracies,
transforming two adjoining rects in this manner can result in the
transformed rects no longer being adjoining. This leads to issues such
as visible seams between tiles when compositing in low quality pinch
zoom mode. This patch fixes this by transforming the min and max
points of the rect rather than its origin and size.

We must ensure we maintain the existing behaviour of returning a
zero-sized rect when the input rect had a negative size, otherwise
certain reftests fail. Eventually we should find the root cause of
these invalid input rects and can change this to an assertion. For now
we just add a test to explicitly ensure this is not regressed.

Differential Revision: https://phabricator.services.mozilla.com/D290832