Handle Android MediaCodec samples being released out-of-order
Usually samples are rendered to the surface in the same order they are
processed, when the corresponding Image becomes current. However, if
we decide to not render a sample we will release it immediately
without rendering it. Since we use a FIFO queue data structure to map
samples to the buffer index to release, this results in us releasing
the wrong sample, then subsequently rendering the wrong samples.

This patch fixes the bug by iterating through the queue of samples
until it finds the matching one, and releasing its buffer. We
additionally inform the caller of releaseOutput() whether the sample
is actually rendered.

Neither of these issues seem to matter much currently, presumably
because on Android we always update the SurfaceTexture to the most
recent frame prior to rendering. They may have been causing an
incorrect frame to be rendered here or there, but nobody seems to have
noticed. Subsequent patches in this series, however, add the ability
to render to an ImageReader rather than a SurfaceTexture, where we
must be more deliberate about when we acquire new images.