Avoid crash due to screenshotting
Need to remove the assertion that the readback format is BGRA, as we
use RGBA on Android.

Can't just return null for CreateScreenshotter, as that makes us fall
back to screenshotting in webrender, which will cause a GL error due
to there being no default framebuffer.

So instead, for now, create a dummy screenshot implementation which
pretends to succeed.