Bug 2028404 - Remove TextureFactoryIdentifier argument from BrowserParent::SendInitRendering. r?#gfx-reviewers
In the process of making WebRenderLayerManager initialisation more
asynchronous, we plan to remove the
WebRenderBridgeChild::SendEnsureConnected() call from
WebRenderLayerManager::Initialize(). The bridge will instead attempt
to connect asynchronously, but if anything attempts to use it prior to
this completing we must fall back to calling SendEnsureConnected()
anyway.

When we remove the call from Initialize(), the first place we will
instead need to synchronously block for connection will be the
RemoteLayerTreeOwner::GetTextureFactoryIdentifier() call in
BrowserParent::InitRendering(). This is called so that the
TextureFactoryIdentifier can be sent to the BrowserChild. However, the
BrowserChild no longer uses the value it receives in
RecvInitRendering(). Instead, it uses a value it obtains from
initialising its own layer manager.

This patch removes the TextureFactoryIdentifier argument from
SendInitRendering(), allowing us to also remove the call to
RemoteLayerTreeOwner::GetTextureFactoryIdentifier() and the function
itself.

Differential Revision: https://phabricator.services.mozilla.com/D291430