Bug 1688963 - Add async initialization function to WebRenderLayerManager
Adds a (currently unused) function
WebRenderLayerManager::InitializeAsync(). This is similar to the
existing Initialize(), but rather than blocking synchronously for the
bridge to connect it returns a promise that resolves when the bridge has
connected (or failed to connect).

Callers can chain to this to perform work once the bridge has
connected. It's not, however, feasible to expect all users of
WebRenderLayerManager to chain onto this promise. We must therefore
still ensure that the bridge has connected prior to anything using
it. To achieve this we make the helper function
WebRenderLayerManager::WrBridge() first call mWrChild->EnsureConnected()
before returning mWrChild.

Note that it is fine if this connection fails: the parent side will be
put into a destroyed state and correctly handle any incoming messages.

Additionally this patch moves some common code from the old Initialize()
function into shared paths: either into the constructor, or by lazily
initializing the display list builder in EndTransactionWithoutLayer().